trace task_io
The "write_bytes" curve over time is very handy for viewing the
smoothness and fairness of page dirties for each task.

XXX: It looks not a good fit for writeback traces, shall we create a
include/trace/events/vfs.h for this?

XXX: the other fields of struct task_io_accounting are not dumped.

Signed-off-by: Wu Fengguang <fengguang.wu@intel.com>